ocr-service/requirements.txt

Flask
google-cloud-vision
